requirements.txt
argcomplete==1.9.3
astroid==2.2.5
boto3==1.9.221
botocore==1.12.221
certifi==2019.6.16
cfn-flip==1.2.1
chardet==3.0.4
Click==7.0
Django==2.2.5
djangorestframework==3.10.2
docutils==0.15.2
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.8
isort==4.3.21
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
lazy-object-proxy==1.4.2
mccabe==0.6.1
placebo==0.9.0
psycopg2-binary==2.8.3
pylint==2.3.1
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.1
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
wsgi-request-logger==0.4.6
zappa==0.48.2
